+++ front/vite-project/src/router/index.js
@@ -1,7 +1,7 @@
 import { createRouter, createWebHistory } from 'vue-router'
 import Content from '../components/Content.vue'
 import Login from '../views/Login.vue'
-import Review from '../views/Review.vue'
+import Review from '../views/review/Review.vue'
 
 const routes = [
   { path: '/', component: Content },
